+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 * Add `dateOfBirth` and `countryCode` to `TransactionIndustryDataRequest`
+* Add `SubscriptionBillingSkipped` webhook notification
+
+## 3.27.0
+* Add validation error codes:
+  * `FAILED_AUTH_ADJUSTMENT_FOR_AMOUNT_SUBMITTED_HARD_DECLINE`
+  * `FAILED_AUTH_ADJUSTMENT_FOR_AMOUNT_SUBMITTED_ALLOW_RETRY`
+* Add `implicitlyVaultedPaymentMethodToken` and `implicitlyVaultedPaymentMethodGlobalId` to `LocalPaymentDetails`
 
 ## 3.26.0
 * Add `arrivalDate`, `ticketIssuerAddress` to `TransactionIndustryRequest`